requirements.txt
Django==4.1.4
djangorestframework==3.14.0
facenet_pytorch==2.5.2
numpy==1.24.1
opencv_python==4.7.0.68
Pillow==9.4.0
six==1.16.0
torch==1.13.1
gunicorn==20.1.0
redis==4.3.5
celery==5.2.7
django-celery-results==2.4.0
django-cors-headers
psycopg2-binary==2.9.5